==> R/05-HHBase.R <==
# 04-HHBase.R
# Builds the base data.table for households
#
# Copyright © 2016: Majid Einian
# Licence: GPL-3

rm(list=ls())

starttime <- proc.time()
cat("\n\n================ HHBase =====================================\n")

library(yaml)
Settings <- yaml.load_file("Settings.yaml")

#library(foreign)
library(data.table)
library(stringr)

# BigD <- data.table(HHID=NA,Region=NA,Year=NA,Quarter=NA,Month=NA,ProvinceCode=NA)[0]


for(year in (Settings$startyear:Settings$endyear)){
  cat(paste0("\n------------------------------\nYear:",year,"\n"))
  
  load(file=paste0(Settings$HEISRawPath,"Y",year,"Raw.rda"))
  
  if(year >86 & year < 92 ){ 
  load(file=paste0(Settings$HEISCountyCodePath,"Y",year,Settings$HEISCountyCodeFileName,".rda"))
  }
  
  if(year < 87){           # RxxData & UxxData tables are provided Since 1387
    RData <- Tables[[paste0("R",year,"P2")]][,1,with=FALSE]
    RData[, Region:=factor(x="Rural",levels=c("Urban","Rural"))]
    UData <- Tables[[paste0("U",year,"P2")]][,1,with=FALSE]
    UData[, Region:=factor(x="Urban",levels=c("Urban","Rural"))]
    HHBase <- rbind(RData, UData)
    rm(RData,UData)
    setnames(HHBase,c("HHID","Region"))
    HHBase[,Year:=year]
    if(year==74)
      HHBase[,HHIDs:=formatC(HHID, width = 8, flag = "0")]
    else if(year<77)
      HHBase[,HHIDs:=formatC(HHID, width = 7, flag = "0")]
    else if(year %in% 77:86)
      HHBase[,HHIDs:=formatC(HHID, width = 9, flag = "0")]
    
    if(year < 77)
      HHBase[,Quarter:=as.integer(str_sub(HHIDs,4,4))]
    else
      HHBase[,Quarter:=as.integer(str_sub(HHIDs,6,6))]
    HHBase[,Month:=NA_integer_]
  }else{
    RData <- Tables[[paste0("R",year,"DATA")]][,c(1:2),with=FALSE]
    RData[, Region:=factor(x="Rural",levels=c("Urban","Rural"))]
    UData <- Tables[[paste0("U",year,"DATA")]][,c(1:2),with=FALSE]
    UData[, Region:=factor(x="Urban",levels=c("Urban","Rural"))]
    HHBase <- rbind(RData, UData)
    rm(RData,UData)
    setnames(HHBase,c("HHID","Month","Region"))
    HHBase[,Month:=ifelse(Month==1,12,Month - 1)]
    HHBase[,Quarter:=(Month-1)%/%3+1]
    HHBase[,HHIDs:=as.character(HHID)]
  }
  
  if(year >86 & year < 92 ){ 
 HHBase<-merge(HHBase,ShCode,by="HHID",all.x = TRUE)
    }
  
  HHBase <- HHBase[!is.na(HHID)]
  HHBase[,ProvinceCode:=as.integer(str_sub(HHIDs,2,3))]
  
  if(year <87 | year > 91 ){
  HHBase[,CountyCode:=as.integer(str_sub(HHIDs,2,5))]
  }
  
  if(year >86 & year < 92 ){ 
    HHBase[,CountyCode:=as.integer(SHCode)]
  }
  
  HHBase[,Year:=year]
  HHBase <- HHBase[,.(HHID,Region,Year,Quarter,Month,ProvinceCode,CountyCode)]
  
  HHBase[,NewArea:=ProvinceCode]
  HHBase[Region=="Urban" & 
           CountyCode %in% c(2301,303,603,707,
                             916,1002,3001,502,2202,
                             401,808,1,1105,
                             1304,2105,105),
         NewArea:=CountyCode]
  
  #Tehran-Alborz
  if(year >76 & year < 92 ){ 
    HHBase[ Region=="Urban" & CountyCode %in% c(2305),
            NewArea:=3001] 
  }
  
  if(year >76 & year < 92 ){ 
    HHBase[CountyCode %in% c(2308,2315,2309),
           NewArea:=30]
  }
  
  #Khorasan
  if(year >76 & year < 87 ){ 
    HHBase[Region=="Urban" & CountyCode %in% c(916),
           NewArea:=916]
  }
  
  if(year >76 & year < 87 ){ 
    HHBase[CountyCode %in% c(901,902,909,924,925),
           NewArea:=28]
  }
  
  if(year >76 & year < 87 ){ 
    HHBase[CountyCode %in% c(903,911,912,921),
           NewArea:=29]
  }
  
  #Ghazvin
  if(year >76 & year < 82 ){ 
    HHBase[CountyCode %in% c(2311),
           NewArea:=26]
  }
  
  #Golestan
  if(year >76 & year < 82 ){ 
    HHBase[CountyCode %in% c(212,203,209,211,213,217),
           NewArea:=27]
  }
  
  
  HHBase[,NewArea2:=as.factor(NewArea)]
  
  HHBase[NewArea==0,NewArea2:="Markazi"]
  HHBase[NewArea==1,NewArea2:="Gilan"]
  HHBase[NewArea==2,NewArea2:="Mazandaran"]
  HHBase[NewArea==3,NewArea2:="Az_Sharghi"]
  HHBase[NewArea==4,NewArea2:="Az_Gharbi"]
  HHBase[NewArea==5,NewArea2:="Kermanshah"]
  HHBase[NewArea==6,NewArea2:="Khoozestan"]
  HHBase[NewArea==7,NewArea2:="Fars"]
  HHBase[NewArea==8,NewArea2:="Kerman"]
  HHBase[NewArea==9,NewArea2:="Khorasan_Razavi"]
  HHBase[NewArea==10,NewArea2:="Esfahan"]
  HHBase[NewArea==11,NewArea2:="Sistan"]
  HHBase[NewArea==12,NewArea2:="Kordestan"]
  HHBase[NewArea==13,NewArea2:="Hamedan"]
  HHBase[NewArea==14,NewArea2:="Chaharmahal"]
  HHBase[NewArea==15,NewArea2:="Lorestan"]
  HHBase[NewArea==16,NewArea2:="Ilam"]
  HHBase[NewArea==17,NewArea2:="Kohkilooye"]
  HHBase[NewArea==18,NewArea2:="Booshehr"]
  HHBase[NewArea==19,NewArea2:="Zanjan"]
  HHBase[NewArea==20,NewArea2:="Semnan"]
  HHBase[NewArea==21,NewArea2:="Yazd"]
  HHBase[NewArea==22,NewArea2:="Hormozgan"]
  HHBase[NewArea==23,NewArea2:="Tehran"]
  HHBase[NewArea==24,NewArea2:="Ardebil"]
  HHBase[NewArea==25,NewArea2:="Ghom"]
  HHBase[NewArea==26,NewArea2:="Ghazvin"]
  HHBase[NewArea==27,NewArea2:="Golestan"]
  HHBase[NewArea==28,NewArea2:="Khorasan_Shomali"]
  HHBase[NewArea==29,NewArea2:="Khorasan_Jonoobi"]
  HHBase[NewArea==30,NewArea2:="Alborz"]
  HHBase[NewArea==2301,NewArea2:="Sh_Tehran"]
  HHBase[NewArea==303,NewArea2:="Sh_Tabriz"]
  HHBase[NewArea==603,NewArea2:="Sh_Ahvaz"]
  HHBase[NewArea==707,NewArea2:="Sh_Shiraz"]
  HHBase[NewArea==916,NewArea2:="Sh_Mashhad"]
  HHBase[NewArea==1002,NewArea2:="Sh_Esfahan"]
  HHBase[NewArea==3001,NewArea2:="Sh_Karaj"]
  HHBase[NewArea==502,NewArea2:="Sh_Kermanshah"]
  HHBase[NewArea==2202,NewArea2:="Sh_Bandarabas"]
  HHBase[NewArea==401,NewArea2:="Sh_Urmia"]
  HHBase[NewArea==105,NewArea2:="Sh_Rasht"]
  HHBase[NewArea==1105,NewArea2:="Sh_Zahedan"]
  HHBase[NewArea==808,NewArea2:="Sh_Kerman"]
  HHBase[NewArea==1304,NewArea2:="Sh_Hamedan"]
  HHBase[NewArea==2105,NewArea2:="Sh_Yazd"]
  HHBase[NewArea==1 & CountyCode==1 ,NewArea2:="Sh_Arak"]

  HHBase[CountyCode==	0001	,County_Name:="	Arak	"]
  HHBase[CountyCode==	0002	,County_Name:="	Ashtian	"]
  HHBase[CountyCode==	0003	,County_Name:="	Tafresh	"]
  HHBase[CountyCode==	0004	,County_Name:="	Khomein	"]
  HHBase[CountyCode==	0005	,County_Name:="	Delijan	"]
  HHBase[CountyCode==	0006	,County_Name:="	Save	"]
  HHBase[CountyCode==	0007	,County_Name:="	Shazand	"]
  HHBase[CountyCode==	0009	,County_Name:="	Mahallat	"]
  HHBase[CountyCode==	0010	,County_Name:="	Zarandie	"]
  HHBase[CountyCode==	0011	,County_Name:="	Komeyjan	"]
  HHBase[CountyCode==	0012	,County_Name:="	Khandab	"]
  HHBase[CountyCode==	0013	,County_Name:="	Farahan	"]
  HHBase[CountyCode==	0101	,County_Name:="	Astara	"]
  HHBase[CountyCode==	0102	,County_Name:="	stane_Ashrafie	"]
  HHBase[CountyCode==	0103	,County_Name:="	Anzali	"]
  HHBase[CountyCode==	0104	,County_Name:="	Talesh	"]
  HHBase[CountyCode==	0105	,County_Name:="	Rasht	"]
  HHBase[CountyCode==	0106	,County_Name:="	Roodbar	"]
  HHBase[CountyCode==	0107	,County_Name:="	Roodsar	"]
  HHBase[CountyCode==	0108	,County_Name:="	Somee_Sara	"]
  HHBase[CountyCode==	0109	,County_Name:="	Fooman	"]
  HHBase[CountyCode==	0110	,County_Name:="	Langerood	"]
  HHBase[CountyCode==	0111	,County_Name:="	Lahijan	"]
  HHBase[CountyCode==	0112	,County_Name:="	Shaft	"]
  HHBase[CountyCode==	0113	,County_Name:="	Emlash	"]
  HHBase[CountyCode==	0114	,County_Name:="	Rezvanshahr	"]
  HHBase[CountyCode==	0115	,County_Name:="	Siahkal	"]
  HHBase[CountyCode==	0116	,County_Name:="	Masal	"]
  HHBase[CountyCode==	0201	,County_Name:="	Amol	"]
  HHBase[CountyCode==	0202	,County_Name:="	Babol	"]
  HHBase[CountyCode==	0204	,County_Name:="	Behshahr	"]
  HHBase[CountyCode==	0205	,County_Name:="	Tonekabon	"]
  HHBase[CountyCode==	0206	,County_Name:="	Ramsar	"]
  HHBase[CountyCode==	0207	,County_Name:="	Sari	"]
  HHBase[CountyCode==	0208	,County_Name:="	Savadkooh	"]
  HHBase[CountyCode==	0210	,County_Name:="	Ghaemshahr	"]
  HHBase[CountyCode==	0214	,County_Name:="	Noor	"]
  HHBase[CountyCode==	0215	,County_Name:="	Noshahr	"]
  HHBase[CountyCode==	0216	,County_Name:="	Babolsar	"]
  HHBase[CountyCode==	0218	,County_Name:="	Mahmoodabad	"]
  HHBase[CountyCode==	0219	,County_Name:="	Neka	"]
  HHBase[CountyCode==	0220	,County_Name:="	Chaloos	"]
  HHBase[CountyCode==	0221	,County_Name:="	Jooybar	"]
  HHBase[CountyCode==	0222	,County_Name:="	Galoogah	"]
  HHBase[CountyCode==	0223	,County_Name:="	Fereydoonkenar	"]
  HHBase[CountyCode==	0224	,County_Name:="	Abbasabad	"]
  HHBase[CountyCode==	0225	,County_Name:="	Miandorood	"]
  HHBase[CountyCode==	0302	,County_Name:="	Ahar	"]
  HHBase[CountyCode==	0303	,County_Name:="	Tabriz	"]
  HHBase[CountyCode==	0305	,County_Name:="	Sarab	"]
  HHBase[CountyCode==	0306	,County_Name:="	Maraghe	"]
  HHBase[CountyCode==	0307	,County_Name:="	Marand	"]
  HHBase[CountyCode==	0310	,County_Name:="	Miane	"]
  HHBase[CountyCode==	0311	,County_Name:="	Hashtrood	"]
  HHBase[CountyCode==	0312	,County_Name:="	Bonab	"]
  HHBase[CountyCode==	0313	,County_Name:="	Bostanaad	"]
  HHBase[CountyCode==	0314	,County_Name:="	Shabestar	"]
  HHBase[CountyCode==	0315	,County_Name:="	Koleybar	"]
  HHBase[CountyCode==	0316	,County_Name:="	Haris	"]
  HHBase[CountyCode==	0319	,County_Name:="	Jolfa	"]
  HHBase[CountyCode==	0320	,County_Name:="	Malekan	"]
  HHBase[CountyCode==	0321	,County_Name:="	Azarshahr	"]
  HHBase[CountyCode==	0322	,County_Name:="	Oskoo	"]
  HHBase[CountyCode==	0323	,County_Name:="	Charooymagh	"]
  HHBase[CountyCode==	0324	,County_Name:="	Varzaghan	"]
  HHBase[CountyCode==	0325	,County_Name:="	Ajabshir	"]
  HHBase[CountyCode==	0326	,County_Name:="	Khosa_Afarin	"]
  HHBase[CountyCode==	0401	,County_Name:="	Urmia	"]
  HHBase[CountyCode==	0402	,County_Name:="	Piranshahr	"]
  HHBase[CountyCode==	0403	,County_Name:="	Khoy	"]
  HHBase[CountyCode==	0404	,County_Name:="	Sardasht	"]
  HHBase[CountyCode==	0405	,County_Name:="	Salmas	"]
  HHBase[CountyCode==	0406	,County_Name:="	Makoo	"]
  HHBase[CountyCode==	0407	,County_Name:="	Mahabad	"]
  HHBase[CountyCode==	0408	,County_Name:="	Miandoab	"]
  HHBase[CountyCode==	0409	,County_Name:="	Naghade	"]
  HHBase[CountyCode==	0410	,County_Name:="	Bookan	"]
  HHBase[CountyCode==	0411	,County_Name:="	Shahindej	"]
  HHBase[CountyCode==	0412	,County_Name:="	Takab	"]
  HHBase[CountyCode==	0413	,County_Name:="	Oshnavie	"]
  HHBase[CountyCode==	0414	,County_Name:="	Chaldoran	"]
  HHBase[CountyCode==	0415	,County_Name:="	Poldasht	"]
  HHBase[CountyCode==	0416	,County_Name:="	Chaipare	"]
  HHBase[CountyCode==	0417	,County_Name:="	Shoot	"]
  HHBase[CountyCode==	0501	,County_Name:="	Eslamabad_Gharb	"]
  HHBase[CountyCode==	0502	,County_Name:="	Kermanshah	"]
  HHBase[CountyCode==	0503	,County_Name:="	Paveh	"]
  HHBase[CountyCode==	0504	,County_Name:="	Sar_Pol_Zahab	"]
  HHBase[CountyCode==	0505	,County_Name:="	Sangher	"]
  HHBase[CountyCode==	0506	,County_Name:="	Ghasr_Shirin	"]
  HHBase[CountyCode==	0507	,County_Name:="	Kangavar	"]
  HHBase[CountyCode==	0508	,County_Name:="	Gilan_Gharb	"]
  HHBase[CountyCode==	0509	,County_Name:="	Javanrood	"]
  HHBase[CountyCode==	0510	,County_Name:="	Sahne	"]
  HHBase[CountyCode==	0511	,County_Name:="	Hersin	"]
  HHBase[CountyCode==	0512	,County_Name:="	Solas_Babajani	"]
  HHBase[CountyCode==	0513	,County_Name:="	Dalahoo	"]
  HHBase[CountyCode==	0514	,County_Name:="	Ravansar	"]
  HHBase[CountyCode==	0601	,County_Name:="	Abadan	"]
  HHBase[CountyCode==	0602	,County_Name:="	Andimeshk	"]
  HHBase[CountyCode==	0603	,County_Name:="	Ahvaz	"]
  HHBase[CountyCode==	0604	,County_Name:="	Izeh	"]
  HHBase[CountyCode==	0605	,County_Name:="	Bandar_Mahshahr	"]
  HHBase[CountyCode==	0606	,County_Name:="	Behbahan	"]
  HHBase[CountyCode==	0607	,County_Name:="	Khoramshahr	"]
  HHBase[CountyCode==	0608	,County_Name:="	Dezfool	"]
  HHBase[CountyCode==	0609	,County_Name:="	Dasht_Azadegan	"]
  HHBase[CountyCode==	0610	,County_Name:="	Ramehramz	"]
  HHBase[CountyCode==	0611	,County_Name:="	Shadegan	"]
  HHBase[CountyCode==	0612	,County_Name:="	Shooshtar	"]
  HHBase[CountyCode==	0613	,County_Name:="	Masjed_Soleyman	"]
  HHBase[CountyCode==	0614	,County_Name:="	Shooshtar	"]
  HHBase[CountyCode==	0615	,County_Name:="	Bagh_Malek	"]
  HHBase[CountyCode==	0616	,County_Name:="	Omidie	"]
  HHBase[CountyCode==	0617	,County_Name:="	Lali	"]
  HHBase[CountyCode==	0618	,County_Name:="	Handijan	"]
  HHBase[CountyCode==	0619	,County_Name:="	Ramshir	"]
  HHBase[CountyCode==	0620	,County_Name:="	Gatvand	"]
  HHBase[CountyCode==	0621	,County_Name:="	Andika	"]
  HHBase[CountyCode==	0622	,County_Name:="	Haftgol	"]
  HHBase[CountyCode==	0623	,County_Name:="	Hoveyze	"]
  HHBase[CountyCode==	0624	,County_Name:="	Bavi	"]
  HHBase[CountyCode==	0701	,County_Name:="	Abadeh	"]
  HHBase[CountyCode==	0702	,County_Name:="	Estahban	"]
  HHBase[CountyCode==	0703	,County_Name:="	Eghlid	"]
  HHBase[CountyCode==	0704	,County_Name:="	Jahrom	"]
  HHBase[CountyCode==	0705	,County_Name:="	Darab	"]
  HHBase[CountyCode==	0706	,County_Name:="	Sepidan	"]
  HHBase[CountyCode==	0707	,County_Name:="	Shiraz	"]
  HHBase[CountyCode==	0708	,County_Name:="	Fasa	"]
  HHBase[CountyCode==	0709	,County_Name:="	Firoozabad	"]
  HHBase[CountyCode==	0710	,County_Name:="	Kazeroon	"]
  HHBase[CountyCode==	0711	,County_Name:="	Larestan	"]
  HHBase[CountyCode==	0712	,County_Name:="	Marvdasht	"]
  HHBase[CountyCode==	0713	,County_Name:="	Mamsani	"]
  HHBase[CountyCode==	0714	,County_Name:="	Neyriz	"]
  HHBase[CountyCode==	0715	,County_Name:="	Lamerd	"]
  HHBase[CountyCode==	0716	,County_Name:="	Boanat	"]
  HHBase[CountyCode==	0717	,County_Name:="	Arsanjan	"]
  HHBase[CountyCode==	0718	,County_Name:="	Khorambid	"]
  HHBase[CountyCode==	0719	,County_Name:="	Zarindasht	"]
  HHBase[CountyCode==	0720	,County_Name:="	Ghir_Karzin	"]
  HHBase[CountyCode==	0721	,County_Name:="	Mehr	"]
  HHBase[CountyCode==	0722	,County_Name:="	Farashband	"]
  HHBase[CountyCode==	0723	,County_Name:="	Pasargad	"]
  HHBase[CountyCode==	0724	,County_Name:="	Khanaj	"]
  HHBase[CountyCode==	0725	,County_Name:="	Sarvestan	"]
  HHBase[CountyCode==	0726	,County_Name:="	Rostam	"]
  HHBase[CountyCode==	0727	,County_Name:="	Gerash	"]
  HHBase[CountyCode==	0728	,County_Name:="	Kovar	"]
  HHBase[CountyCode==	0729	,County_Name:="	Kharame	"]
  HHBase[CountyCode==	0801	,County_Name:="	Baft	"]
  HHBase[CountyCode==	0802	,County_Name:="	Bam	"]
  HHBase[CountyCode==	0803	,County_Name:="	Jiroft	"]
  HHBase[CountyCode==	0804	,County_Name:="	Rafsanjan	"]
  HHBase[CountyCode==	0805	,County_Name:="	Zarandie	"]
  HHBase[CountyCode==	0806	,County_Name:="	Sirjan	"]
  HHBase[CountyCode==	0807	,County_Name:="	Shahr_Babak	"]
  HHBase[CountyCode==	0808	,County_Name:="	Kerman	"]
  HHBase[CountyCode==	0809	,County_Name:="	Kohnooj	"]
  HHBase[CountyCode==	0810	,County_Name:="	Bardsir	"]
  HHBase[CountyCode==	0811	,County_Name:="	Ravar	"]
  HHBase[CountyCode==	0812	,County_Name:="	Anbarabad	"]
  HHBase[CountyCode==	0813	,County_Name:="	Manojan	"]
  HHBase[CountyCode==	0814	,County_Name:="	Koohbanan	"]
  HHBase[CountyCode==	0815	,County_Name:="	Roodbar_Jonoob	"]
  HHBase[CountyCode==	0816	,County_Name:="	Ghale_Ganj	"]
  HHBase[CountyCode==	0817	,County_Name:="	Rigan	"]
  HHBase[CountyCode==	0818	,County_Name:="	Rabar	"]
  HHBase[CountyCode==	0819	,County_Name:="	Fahrej	"]
  HHBase[CountyCode==	0820	,County_Name:="	Anar	"]
  HHBase[CountyCode==	0821	,County_Name:="	Normashir	"]
  HHBase[CountyCode==	0822	,County_Name:="	Faryab	"]
  HHBase[CountyCode==	0823	,County_Name:="	Arzooie	"]
  HHBase[CountyCode==	0904	,County_Name:="	Taybad	"]
  HHBase[CountyCode==	0905	,County_Name:="	Torbat_Heydarie	"]
  HHBase[CountyCode==	0906	,County_Name:="	Torbat_Jam	"]
  HHBase[CountyCode==	0907	,County_Name:="	Dargez	"]
  HHBase[CountyCode==	0908	,County_Name:="	Sabzevar	"]
  HHBase[CountyCode==	0913	,County_Name:="	Ghoochan	"]
  HHBase[CountyCode==	0914	,County_Name:="	Kashmar	"]
  HHBase[CountyCode==	0915	,County_Name:="	Gonabad	"]
  HHBase[CountyCode==	0916	,County_Name:="	Mashhad	"]
  HHBase[CountyCode==	0917	,County_Name:="	Neishaboor	"]
  HHBase[CountyCode==	0918	,County_Name:="	Chenaran	"]
  HHBase[CountyCode==	0919	,County_Name:="	Khaf	"]
  HHBase[CountyCode==	0920	,County_Name:="	Sarakhs	"]
  HHBase[CountyCode==	0922	,County_Name:="	Fariman	"]
  HHBase[CountyCode==	0923	,County_Name:="	Bardeskan	"]
  HHBase[CountyCode==	0927	,County_Name:="	Roshtkhar	"]
  HHBase[CountyCode==	0928	,County_Name:="	Kalat	"]
  HHBase[CountyCode==	0929	,County_Name:="	Khalilabad	"]
  HHBase[CountyCode==	0930	,County_Name:="	Mehvalat	"]
  HHBase[CountyCode==	0931	,County_Name:="	Bojestan	"]
  HHBase[CountyCode==	0932	,County_Name:="	Binallod	"]
  HHBase[CountyCode==	0933	,County_Name:="	Firooze	"]
  HHBase[CountyCode==	0934	,County_Name:="	Jaghtay	"]
  HHBase[CountyCode==	0935	,County_Name:="	Zave	"]
  HHBase[CountyCode==	0936	,County_Name:="	Jovin	"]
  HHBase[CountyCode==	0937	,County_Name:="	Bakherz	"]
  HHBase[CountyCode==	0938	,County_Name:="	Khooshab	"]
  HHBase[CountyCode==	1001	,County_Name:="	Ardestan	"]
  HHBase[CountyCode==	1002	,County_Name:="	Esfahan	"]
  HHBase[CountyCode==	1003	,County_Name:="	Khomeinishahr	"]
  HHBase[CountyCode==	1004	,County_Name:="	Khansar	"]
  HHBase[CountyCode==	1005	,County_Name:="	Samirom	"]
  HHBase[CountyCode==	1006	,County_Name:="	Faridan	"]
  HHBase[CountyCode==	1007	,County_Name:="	Fereydoonshahr	"]
  HHBase[CountyCode==	1008	,County_Name:="	Falaverjan	"]
  HHBase[CountyCode==	1009	,County_Name:="	Shahr_Reza	"]
  HHBase[CountyCode==	1010	,County_Name:="	Kashan	"]
  HHBase[CountyCode==	1011	,County_Name:="	Golpaygan	"]
  HHBase[CountyCode==	1012	,County_Name:="	Lanjan	"]
  HHBase[CountyCode==	1013	,County_Name:="	Nayin	"]
  HHBase[CountyCode==	1014	,County_Name:="	Najafaad	"]
  HHBase[CountyCode==	1015	,County_Name:="	Natanz	"]
  HHBase[CountyCode==	1016	,County_Name:="	Shahinshahr	"]
  HHBase[CountyCode==	1017	,County_Name:="	Mobarake	"]
  HHBase[CountyCode==	1018	,County_Name:="	Aran_Bidgol	"]
  HHBase[CountyCode==	1019	,County_Name:="	Tiran	"]
  HHBase[CountyCode==	1020	,County_Name:="	Chadegan	"]
  HHBase[CountyCode==	1021	,County_Name:="	Dehaghan	"]
  HHBase[CountyCode==	1022	,County_Name:="	Barkhar	"]
  HHBase[CountyCode==	1023	,County_Name:="	Khor_Biabanak	"]
  HHBase[CountyCode==	1101	,County_Name:="	Iranshahr	"]
  HHBase[CountyCode==	1102	,County_Name:="	Chahbahar	"]
  HHBase[CountyCode==	1103	,County_Name:="	Khash	"]
  HHBase[CountyCode==	1104	,County_Name:="	Zabol	"]
  HHBase[CountyCode==	1105	,County_Name:="	Zahedan	"]
  HHBase[CountyCode==	1106	,County_Name:="	Saravan	"]
  HHBase[CountyCode==	1107	,County_Name:="	Nikshahr	"]
  HHBase[CountyCode==	1108	,County_Name:="	Sarbaz	"]
  HHBase[CountyCode==	1109	,County_Name:="	Kenarak	"]
  HHBase[CountyCode==	1110	,County_Name:="	Zahak	"]
  HHBase[CountyCode==	1111	,County_Name:="	Hirmand	"]
  HHBase[CountyCode==	1112	,County_Name:="	Delgan	"]
  HHBase[CountyCode==	1113	,County_Name:="	Mehrestan	"]
  HHBase[CountyCode==	1114	,County_Name:="	Sib_Soran	"]
  HHBase[CountyCode==	1201	,County_Name:="	Bane	"]
  HHBase[CountyCode==	1202	,County_Name:="	Bijar	"]
  HHBase[CountyCode==	1203	,County_Name:="	Saghez	"]
  HHBase[CountyCode==	1204	,County_Name:="	Sanandaj	"]
  HHBase[CountyCode==	1205	,County_Name:="	Ghorve	"]
  HHBase[CountyCode==	1206	,County_Name:="	Marivan	"]
  HHBase[CountyCode==	1207	,County_Name:="	Divandare	"]
  HHBase[CountyCode==	1208	,County_Name:="	Kamiaran	"]
  HHBase[CountyCode==	1209	,County_Name:="	Sarvabad	"]
  HHBase[CountyCode==	1210	,County_Name:="	Dehlagan	"]
  HHBase[CountyCode==	1301	,County_Name:="	Tooyserkan	"]
  HHBase[CountyCode==	1302	,County_Name:="	Malayer	"]
  HHBase[CountyCode==	1303	,County_Name:="	Nahavand	"]
  HHBase[CountyCode==	1304	,County_Name:="	Hamedan	"]
  HHBase[CountyCode==	1305	,County_Name:="	Kabootar_Ahang	"]
  HHBase[CountyCode==	1306	,County_Name:="	Asad_Abad	"]
  HHBase[CountyCode==	1307	,County_Name:="	Bahar	"]
  HHBase[CountyCode==	1308	,County_Name:="	Razan	"]
  HHBase[CountyCode==	1309	,County_Name:="	Famenin	"]
  HHBase[CountyCode==	1401	,County_Name:="	Boroojen	"]
  HHBase[CountyCode==	1402	,County_Name:="	Shahrekord	"]
  HHBase[CountyCode==	1403	,County_Name:="	Farsan	"]
  HHBase[CountyCode==	1404	,County_Name:="	Lordegan	"]
  HHBase[CountyCode==	1405	,County_Name:="	Ardal	"]
  HHBase[CountyCode==	1406	,County_Name:="	Koohrang	"]
  HHBase[CountyCode==	1407	,County_Name:="	Kiar	"]
  HHBase[CountyCode==	1501	,County_Name:="	Aligoodarz	"]
  HHBase[CountyCode==	1502	,County_Name:="	Boroojerd	"]
  HHBase[CountyCode==	1503	,County_Name:="	Khoramabad	"]
  HHBase[CountyCode==	1504	,County_Name:="	Dolfan	"]
  HHBase[CountyCode==	1505	,County_Name:="	Dorood	"]
  HHBase[CountyCode==	1506	,County_Name:="	Koohdasht	"]
  HHBase[CountyCode==	1507	,County_Name:="	Azna	"]
  HHBase[CountyCode==	1508	,County_Name:="	Poldokhtar	"]
  HHBase[CountyCode==	1509	,County_Name:="	Selsele	"]
  HHBase[CountyCode==	1510	,County_Name:="	Dore	"]
  HHBase[CountyCode==	1601	,County_Name:="	Ilam	"]
  HHBase[CountyCode==	1602	,County_Name:="	Dareshahr	"]
  HHBase[CountyCode==	1603	,County_Name:="	Dehloran	"]
  HHBase[CountyCode==	1604	,County_Name:="	Cherdavel	"]
  HHBase[CountyCode==	1605	,County_Name:="	Mehran	"]
  HHBase[CountyCode==	1606	,County_Name:="	Abdanan	"]
  HHBase[CountyCode==	1607	,County_Name:="	Ivan	"]
  HHBase[CountyCode==	1608	,County_Name:="	Malekshahi	"]
  HHBase[CountyCode==	1701	,County_Name:="	Boyerahmad	"]
  HHBase[CountyCode==	1702	,County_Name:="	Kohkilooye	"]
  HHBase[CountyCode==	1703	,County_Name:="	Gachsaran	"]
  HHBase[CountyCode==	1704	,County_Name:="	Dena	"]
  HHBase[CountyCode==	1705	,County_Name:="	Bambai	"]
  HHBase[CountyCode==	1706	,County_Name:="	Charam	"]
  HHBase[CountyCode==	1707	,County_Name:="	Basht	"]
  HHBase[CountyCode==	1801	,County_Name:="	Booshehr	"]
  HHBase[CountyCode==	1802	,County_Name:="	Tangestan	"]
  HHBase[CountyCode==	1803	,County_Name:="	Dashtestan	"]
  HHBase[CountyCode==	1804	,County_Name:="	Dashti	"]
  HHBase[CountyCode==	1805	,County_Name:="	Dayer	"]
  HHBase[CountyCode==	1806	,County_Name:="	Kangan	"]
  HHBase[CountyCode==	1807	,County_Name:="	Gonave	"]
  HHBase[CountyCode==	1808	,County_Name:="	Deylam	"]
  HHBase[CountyCode==	1809	,County_Name:="	Jam	"]
  HHBase[CountyCode==	1901	,County_Name:="	Abhar	"]
  HHBase[CountyCode==	1903	,County_Name:="	Khodabande	"]
  HHBase[CountyCode==	1904	,County_Name:="	Zanjan	"]
  HHBase[CountyCode==	1906	,County_Name:="	Ijrood	"]
  HHBase[CountyCode==	1907	,County_Name:="	Khoramdare	"]
  HHBase[CountyCode==	1908	,County_Name:="	Tarem	"]
  HHBase[CountyCode==	1909	,County_Name:="	Mahneshan	"]
  HHBase[CountyCode==	2001	,County_Name:="	Damghan	"]
  HHBase[CountyCode==	2002	,County_Name:="	Semnan	"]
  HHBase[CountyCode==	2003	,County_Name:="	Shahrood	"]
  HHBase[CountyCode==	2004	,County_Name:="	Garmsar	"]
  HHBase[CountyCode==	2005	,County_Name:="	Mehdishahr	"]
  HHBase[CountyCode==	2101	,County_Name:="	Ardakan	"]
  HHBase[CountyCode==	2102	,County_Name:="	Bafgh	"]
  HHBase[CountyCode==	2103	,County_Name:="	Taft	"]
  HHBase[CountyCode==	2104	,County_Name:="	Mehriz	"]
  HHBase[CountyCode==	2105	,County_Name:="	Yazd	"]
  HHBase[CountyCode==	2106	,County_Name:="	Meybod	"]
  HHBase[CountyCode==	2107	,County_Name:="	Abarkooh	"]
  HHBase[CountyCode==	2108	,County_Name:="	Sedoogh	"]
  HHBase[CountyCode==	2109	,County_Name:="	Khatam	"]
  HHBase[CountyCode==	2110	,County_Name:="	Tabas	"]
  HHBase[CountyCode==	2111	,County_Name:="	Bahabad	"]
  HHBase[CountyCode==	2201	,County_Name:="	Aboomoosa	"]
  HHBase[CountyCode==	2202	,County_Name:="	Bandarabas	"]
  HHBase[CountyCode==	2203	,County_Name:="	Bandarlenge	"]
  HHBase[CountyCode==	2204	,County_Name:="	gheshm	"]
  HHBase[CountyCode==	2205	,County_Name:="	Minab	"]
  HHBase[CountyCode==	2206	,County_Name:="	Jask	"]
  HHBase[CountyCode==	2207	,County_Name:="	Roodan	"]
  HHBase[CountyCode==	2208	,County_Name:="	Hajiabad	"]
  HHBase[CountyCode==	2209	,County_Name:="	Bastak	"]
  HHBase[CountyCode==	2210	,County_Name:="	Khamir	"]
  HHBase[CountyCode==	2211	,County_Name:="	Parsian	"]
  HHBase[CountyCode==	2212	,County_Name:="	Sirik	"]
  HHBase[CountyCode==	2213	,County_Name:="	Bashagerd	"]
  HHBase[CountyCode==	2301	,County_Name:="	Tehran	"]
  HHBase[CountyCode==	2302	,County_Name:="	Damavand	"]
  HHBase[CountyCode==	2303	,County_Name:="	Rey	"]
  HHBase[CountyCode==	2304	,County_Name:="	Shemiranat	"]
  HHBase[CountyCode==	2306	,County_Name:="	Varamin	"]
  HHBase[CountyCode==	2309	,County_Name:="	Shahriar	"]
  HHBase[CountyCode==	2310	,County_Name:="	Eslamshahr	"]
  HHBase[CountyCode==	2312	,County_Name:="	Robatkarim	"]
  HHBase[CountyCode==	2313	,County_Name:="	Pakdasht	"]
  HHBase[CountyCode==	2314	,County_Name:="	Firoozkooh	"]
  HHBase[CountyCode==	2316	,County_Name:="	Ghods	"]
  HHBase[CountyCode==	2317	,County_Name:="	Malard	"]
  HHBase[CountyCode==	2318	,County_Name:="	Pishva	"]
  HHBase[CountyCode==	2319	,County_Name:="	Baharestan	"]
  HHBase[CountyCode==	2401	,County_Name:="	Ardebil	"]
  HHBase[CountyCode==	2402	,County_Name:="	Bilesavar	"]
  HHBase[CountyCode==	2403	,County_Name:="	Khalkhal	"]
  HHBase[CountyCode==	2404	,County_Name:="	Meshginshahr	"]
  HHBase[CountyCode==	2405	,County_Name:="	Germi	"]
  HHBase[CountyCode==	2406	,County_Name:="	Parsabad	"]
  HHBase[CountyCode==	2407	,County_Name:="	Kosar	"]
  HHBase[CountyCode==	2408	,County_Name:="	Samin	"]
  HHBase[CountyCode==	2409	,County_Name:="	Nayer	"]
  HHBase[CountyCode==	2410	,County_Name:="	Sareyn	"]
  HHBase[CountyCode==	2501	,County_Name:="	Ghom	"]
  HHBase[CountyCode==	2601	,County_Name:="	Booyinzahra	"]
  HHBase[CountyCode==	2602	,County_Name:="	Takestan	"]
  HHBase[CountyCode==	2603	,County_Name:="	Ghazvin	"]
  HHBase[CountyCode==	2604	,County_Name:="	Abeyk	"]
  HHBase[CountyCode==	2605	,County_Name:="	Alborz	"]
  HHBase[CountyCode==	2701	,County_Name:="	Bandar_Gaz	"]
  HHBase[CountyCode==	2702	,County_Name:="	Torkaman	"]
  HHBase[CountyCode==	2703	,County_Name:="	Aliabad	"]
  HHBase[CountyCode==	2704	,County_Name:="	Kordkooy	"]
  HHBase[CountyCode==	2705	,County_Name:="	Gorgan	"]
  HHBase[CountyCode==	2706	,County_Name:="	Gonbad_Kavoos	"]
  HHBase[CountyCode==	2707	,County_Name:="	Minoodasht	"]
  HHBase[CountyCode==	2708	,County_Name:="	Agh_Ghola	"]
  HHBase[CountyCode==	2709	,County_Name:="	Kalale	"]
  HHBase[CountyCode==	2710	,County_Name:="	Azadshahr	"]
  HHBase[CountyCode==	2711	,County_Name:="	Ramian	"]
  HHBase[CountyCode==	2712	,County_Name:="	Morace_Tape	"]
  HHBase[CountyCode==	2713	,County_Name:="	Gomeyshan	"]
  HHBase[CountyCode==	2714	,County_Name:="	Galiksh	"]
  HHBase[CountyCode==	2801	,County_Name:="	Esfarayen	"]
  HHBase[CountyCode==	2802	,County_Name:="	Bojnoord	"]
  HHBase[CountyCode==	2803	,County_Name:="	Jajrem	"]
  HHBase[CountyCode==	2804	,County_Name:="	Shirvan	"]
  HHBase[CountyCode==	2805	,County_Name:="	Faroj	"]
  HHBase[CountyCode==	2806	,County_Name:="	Mane_Somolghan	"]
  HHBase[CountyCode==	2807	,County_Name:="	Garme	"]
  HHBase[CountyCode==	2901	,County_Name:="	Birjand	"]
  HHBase[CountyCode==	2902	,County_Name:="	Darmian	"]
  HHBase[CountyCode==	2903	,County_Name:="	Sarbishe	"]
  HHBase[CountyCode==	2904	,County_Name:="	Ghayenat	"]
  HHBase[CountyCode==	2905	,County_Name:="	Nahbandan	"]
  HHBase[CountyCode==	2906	,County_Name:="	Sarayan	"]
  HHBase[CountyCode==	2907	,County_Name:="	Ferdos	"]
  HHBase[CountyCode==	2908	,County_Name:="	Bashrooye	"]
  HHBase[CountyCode==	3001	,County_Name:="	Karaj	"]
  HHBase[CountyCode==	3002	,County_Name:="	Savojbolagh	"]
  HHBase[CountyCode==	3003	,County_Name:="	Nazarabad	"]
  HHBase[CountyCode==	3004	,County_Name:="	Taleghan	"]
  

  save(HHBase, file=paste0(Settings$HEISProcessedPath,"Y",year,"HHBase.rda"))

 # rm(HHBase)
}

endtime <- proc.time()

cat("\n\n============================\nIt took ")
cat(endtime-starttime)
